增加异常处理页面，实现阅读数变化

+++ src/main/java/site/okliu/newvision/controller/PublishController.java
@@ -24,7 +24,7 @@ public class PublishController {
     }
 
     @GetMapping("/{id}")
-    public String edit(@PathVariable("id") Long id, Model model) {
+    public String edit(@PathVariable("id") Integer id, Model model) {
         QuestionDTO question = questionService.findById(id);
         model.addAttribute("title", question.getTitle());
         model.addAttribute("description", question.getDescription());
